Moved header implementations to separate directory
