+++ noether_filtering/include/noether_filtering/impl/filter_group.hpp
